fix: SearchParams 사용으로 Suspense 적용
